ai_service_platform/requirements.txt
flask==2.0.1
flask-sock==0.6.0
werkzeug==2.0.3
requests==2.31.0
numpy==1.21.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
torch==2.0.1
python-magic==0.4.27
python-dotenv==1.0.0
huggingface-hub==0.19.4
websockets==11.0.3
jieba==0.42.1
pymilvus==2.3.1
psutil==6.1.1
